basic group reading

+++ TES4Lib/Base/Record.cs
@@ -5,10 +5,10 @@ namespace TES4Lib.Structures.Base
     public class Record
     {
         public string Name { get; set; }
-        public ulong Size { get; set; }
-        public ulong Flag { get; set; }
+        public int Size { get; set; }
+        public int Flag { get; set; }
         public byte[] FormId { get; set; }
-        public ulong VersionControlInfo { get; set; }
+        public int VersionControlInfo { get; set; }
         public byte[] Data { get; set; }
         private byte[] RawData { get; set; }
 
@@ -17,8 +17,8 @@ namespace TES4Lib.Structures.Base
             RawData = rawData;
             var reader = new ByteReader();
             Name = reader.ReadBytes<string>(RawData, 4);
-            Size = reader.ReadBytes<ulong>(RawData);
-            Flag = reader.ReadBytes<ulong>(RawData, 4);
+            Size = reader.ReadBytes<int>(RawData);
+            Flag = reader.ReadBytes<int>(RawData, 4);
             FormId = reader.ReadBytes<byte[]>(RawData, 8);
             Data = reader.ReadBytes<byte[]>(RawData,(int)Size);
         }
